#54 feat : Board            soft delete 구현            N+1문제 해결 (Querydsl fetch join)            member softdelete에 따른 select문에 조건 추가 (member - isDeleted)            Comment            soft delete 구현            N+1문제 해결 (Querydsl fetch join)            member softdelete에 따른 select문에 조건 추가 (member - isDeleted)            Category            N+1문제 해결 (Querydsl fetch join)

+++ src/main/java/kimp/community/dto/board/response/BoardWithCountResponseDto.java
@@ -11,5 +11,5 @@ import java.util.List;
 @NoArgsConstructor
 public class BoardWithCountResponseDto {
     List<BoardResponseDto> boardResponseDtos;
-    Integer count;
+    Long count;
 }
